create CRUD resources for states module

+++ src/countries/countries.service.ts
@@ -3,6 +3,8 @@ import {
   Injectable,
   InternalServerErrorException,
   NotFoundException,
+  Inject,
+  forwardRef,
 } from '@nestjs/common';
 import { InjectModel } from '@nestjs/mongoose';
 import { Model } from 'mongoose';
@@ -11,11 +13,16 @@ import { CreateCountryDto } from './dto/create-country.dto';
 import { UpdateCountryDto } from './dto/update-country.dto';
 import { Country } from './schemas/country.schema';
 import { DUPLICATE_KEY_CODE } from 'src/constants/mongoErrCodes';
+import { State } from 'src/states/schemas/state.schema';
 
 @Injectable()
 export class CountriesService {
   constructor(
-    @InjectModel(Country.name) private countryModel: Model<Country>,
+    @InjectModel(Country.name)
+    @Inject(forwardRef(() => Model<Country>))
+    private countryModel: Model<Country>,
+
+    @InjectModel(State.name) private stateModel: Model<State>,
   ) {}
 
   async create(createCountryDto: CreateCountryDto): Promise<Country> {
@@ -69,6 +76,11 @@ export class CountriesService {
   }
 
   async remove(id: string): Promise<Country> {
+    const state = await this.stateModel.findOne({ countryId: id });
+    if (state) {
+      throw new BadRequestException('Country is being used.');
+    }
+
     const country = await this.countryModel.findByIdAndDelete(id).exec();
 
     if (!country) {
